refactor: simplify raster export layers
