+++ packages/react-sketch-canvas/src/Canvas/export/image.ts
@@ -1,207 +1,14 @@
 import type * as React from "react";
 import type { ExportImageOptions, ExportImageType } from "../../types";
+import {
+	drawBackgroundLayer,
+	loadBackgroundLayer,
+	resolveBackgroundLayerPlan,
+} from "./backgroundLayer";
+import { encodeSvgDataUrl } from "./encoding";
+import { loadImage } from "./imageLoader";
 import { prepareSvgForExport, removeBackgroundImageFromSvg } from "./svg";
 
-/**
- * The resolved image element produced by browser image loading.
- */
-type LoadImageReturns = Promise<HTMLImageElement>;
-
-/**
- * Decide whether canvas export should request anonymous cross-origin image loading.
- *
- * @remarks
- * Data and blob URLs are already local to the document. Only cross-origin HTTP
- * images need `crossorigin="anonymous"` so the exported canvas can remain
- * readable when the remote server sends compatible CORS headers.
- */
-function shouldUseAnonymousCrossOrigin(url: string): boolean {
-	if (url.startsWith("data:") || url.startsWith("blob:")) {
-		return false;
-	}
-
-	try {
-		const parsedUrl = new URL(url, window.location.href);
-
-		if (!/^https?:$/.test(parsedUrl.protocol)) {
-			return false;
-		}
-
-		return parsedUrl.origin !== window.location.origin;
-	} catch {
-		return false;
-	}
-}
-
-/**
- * Decide whether a configured background image must be painted outside the SVG.
- *
- * @remarks
- * Raster export composites the background as a normal canvas image layer for
- * every URL type. Keeping data/blob URLs inside the serialized SVG would make
- * browsers rasterize the SVG pattern themselves, which can tile the background
- * instead of stretching it to the export dimensions. External image URLs also
- * need this path because browsers do not reliably load external resources from
- * an SVG that is itself being rasterized as an image.
- */
-function shouldPaintBackgroundAsRasterLayer(url: string): boolean {
-	return url.length > 0;
-}
-
-function isSvgDataUri(url: string): boolean {
-	return /^data:image\/svg\+xml(?:[;,]|$)/i.test(url);
-}
-
-function decodeSvgDataUri(url: string): string | null {
-	const match = /^data:image\/svg\+xml(?<metadata>[^,]*),(?<data>.*)$/i.exec(
-		url,
-	);
-
-	if (!match?.groups) {
-		return null;
-	}
-
-	const { metadata, data } = match.groups;
-
-	try {
-		if (metadata.includes(";base64")) {
-			return decodeURIComponent(escape(atob(data)));
-		}
-
-		return decodeURIComponent(data);
-	} catch {
-		return null;
-	}
-}
-
-function normalizeSvgDataUriDimensions(url: string): string {
-	const svgText = decodeSvgDataUri(url);
-
-	if (!svgText) {
-		return url;
-	}
-
-	const document = new DOMParser().parseFromString(svgText, "image/svg+xml");
-	const svgElement = document.documentElement;
-
-	if (svgElement.nodeName.toLowerCase() !== "svg") {
-		return url;
-	}
-
-	if (svgElement.hasAttribute("width") && svgElement.hasAttribute("height")) {
-		return url;
-	}
-
-	const viewBox = svgElement.getAttribute("viewBox");
-
-	if (!viewBox) {
-		return url;
-	}
-
-	const [, , width, height] = viewBox
-		.trim()
-		.split(/[\s,]+/)
-		.map(Number);
-
-	if (!Number.isFinite(width) || !Number.isFinite(height)) {
-		return url;
-	}
-
-	if (!svgElement.hasAttribute("width")) {
-		svgElement.setAttribute("width", String(width));
-	}
-
-	if (!svgElement.hasAttribute("height")) {
-		svgElement.setAttribute("height", String(height));
-	}
-
-	const serializedSvg = new XMLSerializer().serializeToString(svgElement);
-
-	return `data:image/svg+xml;charset=utf-8,${encodeURIComponent(serializedSvg)}`;
-}
-
-function createSvgBackgroundLayer({
-	backgroundImage,
-	exportWidth,
-	exportHeight,
-	preserveAspectRatio,
-}: {
-	backgroundImage: string;
-	exportWidth: number;
-	exportHeight: number;
-	preserveAspectRatio: React.SVGAttributes<HTMLImageElement>["preserveAspectRatio"];
-}): SVGElement {
-	const svgNamespace = "http://www.w3.org/2000/svg";
-	const svg = document.createElementNS(svgNamespace, "svg");
-	const defs = document.createElementNS(svgNamespace, "defs");
-	const pattern = document.createElementNS(svgNamespace, "pattern");
-	const image = document.createElementNS(svgNamespace, "image");
-	const rect = document.createElementNS(svgNamespace, "rect");
-
-	svg.setAttribute("xmlns", svgNamespace);
-	svg.setAttribute("width", String(exportWidth));
-	svg.setAttribute("height", String(exportHeight));
-	svg.setAttribute("viewBox", `0 0 ${exportWidth} ${exportHeight}`);
-
-	pattern.setAttribute("id", "react-sketch-canvas-export-background");
-	pattern.setAttribute("x", "0");
-	pattern.setAttribute("y", "0");
-	pattern.setAttribute("width", "100%");
-	pattern.setAttribute("height", "100%");
-	pattern.setAttribute("patternUnits", "userSpaceOnUse");
-
-	image.setAttribute("x", "0");
-	image.setAttribute("y", "0");
-	image.setAttribute("width", "100%");
-	image.setAttribute("height", "100%");
-	image.setAttribute("href", backgroundImage);
-
-	if (preserveAspectRatio) {
-		image.setAttribute("preserveAspectRatio", preserveAspectRatio);
-	}
-
-	rect.setAttribute("x", "0");
-	rect.setAttribute("y", "0");
-	rect.setAttribute("width", "100%");
-	rect.setAttribute("height", "100%");
-	rect.setAttribute("fill", "url(#react-sketch-canvas-export-background)");
-
-	pattern.append(image);
-	defs.append(pattern);
-	svg.append(defs, rect);
-
-	return svg;
-}
-
-/**
- * Load an image URL or data URL for canvas export.
- *
- * @remarks
- * The image is configured for anonymous cross-origin loading so browser canvas
- * security rules allow `toDataURL` when the remote server permits it.
- */
-export const loadImage = (url: string): LoadImageReturns =>
-	new Promise((resolve, reject) => {
-		const img = new Image();
-
-		if (shouldUseAnonymousCrossOrigin(url)) {
-			img.crossOrigin = "anonymous";
-		}
-
-		img.addEventListener("load", () => {
-			if (img.width > 0) {
-				resolve(img);
-				return;
-			}
-			reject(new Error("Image not found"));
-		});
-		img.addEventListener("error", (err) => reject(err));
-		img.src = url;
-	});
-
-/**
- * Arguments required to render a cloned SVG canvas as a raster image.
- */
 type ExportImageFromSvgParams = {
 	id: string;
 	svgCanvas: SVGElement;
@@ -215,131 +22,9 @@ type ExportImageFromSvgParams = {
 	options?: ExportImageOptions;
 };
 
-/**
- * The data URL produced by a raster image export.
- */
 type ExportImageFromSvgReturns = Promise<string>;
 
-type SvgAspectRatioAlign = "Min" | "Mid" | "Max";
-
-type SvgAspectRatioMode = "meet" | "slice";
-
-type ParsedSvgAspectRatio = {
-	xAlign: SvgAspectRatioAlign;
-	yAlign: SvgAspectRatioAlign;
-	mode: SvgAspectRatioMode;
-};
-
-function parseSvgAspectRatio(
-	preserveAspectRatio: React.SVGAttributes<HTMLImageElement>["preserveAspectRatio"],
-): ParsedSvgAspectRatio | null {
-	if (!preserveAspectRatio || preserveAspectRatio === "none") {
-		return null;
-	}
-
-	const [align = "xMidYMid", mode = "meet"] = preserveAspectRatio.split(/\s+/);
-	const alignMatch = /^x(Min|Mid|Max)Y(Min|Mid|Max)$/.exec(align);
-
-	return {
-		xAlign: (alignMatch?.[1] ?? "Mid") as SvgAspectRatioAlign,
-		yAlign: (alignMatch?.[2] ?? "Mid") as SvgAspectRatioAlign,
-		mode: mode === "slice" ? "slice" : "meet",
-	};
-}
-
-function resolveAlignedOffset(
-	containerSize: number,
-	contentSize: number,
-	align: SvgAspectRatioAlign,
-): number {
-	if (align === "Min") {
-		return 0;
-	}
-
-	const remainingSize = containerSize - contentSize;
-
-	if (align === "Max") {
-		return remainingSize;
-	}
-
-	return remainingSize / 2;
-}
-
-function drawBackgroundLayer(
-	context: CanvasRenderingContext2D,
-	backgroundLayer: HTMLImageElement,
-	exportWidth: number,
-	exportHeight: number,
-	preserveAspectRatio: React.SVGAttributes<HTMLImageElement>["preserveAspectRatio"],
-): void {
-	const parsedAspectRatio = parseSvgAspectRatio(preserveAspectRatio);
-	const sourceWidth = backgroundLayer.naturalWidth || backgroundLayer.width;
-	const sourceHeight = backgroundLayer.naturalHeight || backgroundLayer.height;
-
-	if (!parsedAspectRatio || sourceWidth <= 0 || sourceHeight <= 0) {
-		context.drawImage(backgroundLayer, 0, 0, exportWidth, exportHeight);
-		return;
-	}
-
-	const widthScale = exportWidth / sourceWidth;
-	const heightScale = exportHeight / sourceHeight;
-
-	if (parsedAspectRatio.mode === "slice") {
-		const scale = Math.max(widthScale, heightScale);
-		const sourceCropWidth = exportWidth / scale;
-		const sourceCropHeight = exportHeight / scale;
-		const sourceX = resolveAlignedOffset(
-			sourceWidth,
-			sourceCropWidth,
-			parsedAspectRatio.xAlign,
-		);
-		const sourceY = resolveAlignedOffset(
-			sourceHeight,
-			sourceCropHeight,
-			parsedAspectRatio.yAlign,
-		);
-
-		context.drawImage(
-			backgroundLayer,
-			sourceX,
-			sourceY,
-			sourceCropWidth,
-			sourceCropHeight,
-			0,
-			0,
-			exportWidth,
-			exportHeight,
-		);
-		return;
-	}
-
-	const scale = Math.min(widthScale, heightScale);
-	const destinationWidth = sourceWidth * scale;
-	const destinationHeight = sourceHeight * scale;
-	const destinationX = resolveAlignedOffset(
-		exportWidth,
-		destinationWidth,
-		parsedAspectRatio.xAlign,
-	);
-	const destinationY = resolveAlignedOffset(
-		exportHeight,
-		destinationHeight,
-		parsedAspectRatio.yAlign,
-	);
-
-	context.drawImage(
-		backgroundLayer,
-		destinationX,
-		destinationY,
-		destinationWidth,
-		destinationHeight,
-	);
-}
-
-/**
- * Arguments used to prepare the stroke-only SVG layer for raster export.
- */
-type PrepareStrokeSvgForRasterExportArgs = Pick<
+type PrepareStrokeSvgForRasterExportParams = Pick<
 	ExportImageFromSvgParams,
 	| "id"
 	| "svgCanvas"
@@ -357,22 +42,14 @@ type PrepareStrokeSvgForRasterExportArgs = Pick<
  * the raster canvas. This avoids duplicate or browser-specific SVG image
  * artifacts while keeping strokes composited above the background.
  */
-function prepareStrokeSvgForRasterExport(
-	params: PrepareStrokeSvgForRasterExportArgs,
-): SVGElement {
-	const {
-		id,
-		svgCanvas,
-		canvasColor,
-		backgroundImage,
-		exportWithBackgroundImage,
-	} = params;
-
-	if (
-		exportWithBackgroundImage &&
-		backgroundImage &&
-		shouldPaintBackgroundAsRasterLayer(backgroundImage)
-	) {
+function prepareStrokeSvgForRasterExport({
+	id,
+	svgCanvas,
+	canvasColor,
+	backgroundImage,
+	exportWithBackgroundImage,
+}: PrepareStrokeSvgForRasterExportParams): SVGElement {
+	if (exportWithBackgroundImage && backgroundImage) {
 		const strokeOnlySvg = removeBackgroundImageFromSvg(svgCanvas, id);
 
 		return prepareSvgForExport(strokeOnlySvg, {
@@ -389,95 +66,18 @@ function prepareStrokeSvgForRasterExport(
 	});
 }
 
-/**
- * Encode an SVG element as an image data URL that can be loaded into `<canvas>`.
- */
-function encodeSvgDataUrl(svgCanvas: SVGElement): string {
-	const serializedSvg = new XMLSerializer().serializeToString(svgCanvas);
-	const encodedSvg = btoa(unescape(encodeURIComponent(serializedSvg)));
-
-	return `data:image/svg+xml;base64,${encodedSvg}`;
-}
-
-/**
- * Resolve the raster scale used for image export.
- *
- * @remarks
- * Explicit export dimensions are treated as final pixel dimensions. Without an
- * explicit size, the export uses the device pixel ratio so default-size exports
- * retain the visual sharpness of the on-screen canvas.
- */
-function resolveRasterScale(options?: ExportImageOptions): number {
-	const hasExplicitSize =
-		options?.width !== undefined || options?.height !== undefined;
-
-	if (hasExplicitSize) {
-		return 1;
-	}
-
-	return Math.max(window.devicePixelRatio || 1, 1);
-}
-
-/**
- * Create the temporary canvas used to composite background and stroke layers.
- */
 function createRasterCanvas(
 	exportWidth: number,
 	exportHeight: number,
-	pixelRatio: number,
 ): HTMLCanvasElement {
 	const renderCanvas = document.createElement("canvas");
 
-	renderCanvas.width = Math.round(exportWidth * pixelRatio);
-	renderCanvas.height = Math.round(exportHeight * pixelRatio);
+	renderCanvas.width = exportWidth;
+	renderCanvas.height = exportHeight;
 	renderCanvas.style.width = `${exportWidth}px`;
 	renderCanvas.style.height = `${exportHeight}px`;
 
 	return renderCanvas;
-}
-
-/**
- * Load the optional background image layer for raster export.
- *
- * @remarks
- * Background image failures are recoverable: consumers still receive the
- * exported stroke layer, and the warning explains what image/CORS settings to
- * check.
- */
-async function loadBackgroundLayer(
-	backgroundImage: string,
-	exportWithBackgroundImage: boolean,
-	exportWidth: number,
-	exportHeight: number,
-	preserveAspectRatio: React.SVGAttributes<HTMLImageElement>["preserveAspectRatio"],
-): Promise<HTMLImageElement | null> {
-	if (
-		!exportWithBackgroundImage ||
-		!backgroundImage ||
-		!shouldPaintBackgroundAsRasterLayer(backgroundImage)
-	) {
-		return null;
-	}
-
-	try {
-		if (isSvgDataUri(backgroundImage)) {
-			const backgroundSvg = createSvgBackgroundLayer({
-				backgroundImage,
-				exportWidth,
-				exportHeight,
-				preserveAspectRatio,
-			});
-
-			return await loadImage(encodeSvgDataUrl(backgroundSvg));
-		}
-
-		return await loadImage(normalizeSvgDataUriDimensions(backgroundImage));
-	} catch {
-		console.warn(
-			"React Sketch Canvas could not load the background image while exporting. Check that backgroundImage points to a reachable image and allows cross-origin access.",
-		);
-		return null;
-	}
 }
 
 /**
@@ -502,6 +102,13 @@ export async function exportImageFromSvg({
 }: ExportImageFromSvgParams): ExportImageFromSvgReturns {
 	const exportWidth = options?.width ?? svgWidth;
 	const exportHeight = options?.height ?? svgHeight;
+	const backgroundLayerPlan = resolveBackgroundLayerPlan({
+		backgroundImage,
+		exportWithBackgroundImage,
+		exportWidth,
+		exportHeight,
+		preserveAspectRatio: preserveBackgroundImageAspectRatio,
+	});
 	const preparedSvg = prepareStrokeSvgForRasterExport({
 		id,
 		svgCanvas,
@@ -510,20 +117,11 @@ export async function exportImageFromSvg({
 		exportWithBackgroundImage,
 	});
 	const strokeImage = await loadImage(encodeSvgDataUrl(preparedSvg));
-	const pixelRatio = resolveRasterScale(options);
-	const renderCanvas = createRasterCanvas(
-		exportWidth,
-		exportHeight,
-		pixelRatio,
-	);
+	const renderCanvas = createRasterCanvas(exportWidth, exportHeight);
 	const context = renderCanvas.getContext("2d");
 
 	if (!context) {
 		throw Error("Canvas not rendered yet");
-	}
-
-	if (pixelRatio !== 1) {
-		context.scale(pixelRatio, pixelRatio);
 	}
 
 	const shouldFillCanvasBackground =
@@ -534,26 +132,16 @@ export async function exportImageFromSvg({
 		context.fillRect(0, 0, exportWidth, exportHeight);
 	}
 
-	const backgroundLayer = await loadBackgroundLayer(
-		backgroundImage,
-		exportWithBackgroundImage,
-		exportWidth,
-		exportHeight,
-		preserveBackgroundImageAspectRatio,
-	);
+	const backgroundLayer = await loadBackgroundLayer(backgroundLayerPlan);
 
-	if (backgroundLayer) {
-		if (isSvgDataUri(backgroundImage)) {
-			context.drawImage(backgroundLayer, 0, 0, exportWidth, exportHeight);
-		} else {
-			drawBackgroundLayer(
-				context,
-				backgroundLayer,
-				exportWidth,
-				exportHeight,
-				preserveBackgroundImageAspectRatio,
-			);
-		}
+	if (backgroundLayer && backgroundLayerPlan.kind !== "none") {
+		drawBackgroundLayer({
+			context,
+			backgroundLayer,
+			plan: backgroundLayerPlan,
+			exportWidth,
+			exportHeight,
+		});
 	}
 
 	context.drawImage(strokeImage, 0, 0, exportWidth, exportHeight);
